backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic[email]==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.19
requests==2.32.3
httpx==0.27.2
slowapi==0.1.9
alembic==1.14.1
google-generativeai==0.8.4
anthropic==0.40.0
openai>=1.14.0
